api/requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, pytest
backcall==0.1.0           # via ipython
black==19.10b0            # via -r requirements-dev.in
boto3==1.12.18            # via -r requirements.in
botocore==1.15.18         # via boto3, s3transfer
brotlipy==0.7.0           # via -r requirements.in
cffi==1.14.0              # via brotlipy
click==7.1.1              # via black
coverage==5.0.3           # via pytest-cov
datetime==4.3             # via -r requirements-dev.in
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via odfpy
diff-match-patch==20181111  # via django-import-export
dj-database-url==0.5.0    # via -r requirements.in
django-ckeditor==5.9.0    # via -r requirements.in
django-configurations==2.2  # via -r requirements.in
django-import-export==2.0.2  # via -r requirements.in
django-js-asset==1.2.2    # via django-ckeditor
django-storages==1.9.1    # via -r requirements.in
django-taggit==1.2.0      # via -r requirements.in
django==2.2.11            # via -r requirements.in, django-import-export, django-storages, django-taggit, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
docutils==0.15.2          # via botocore
entrypoints==0.3          # via flake8
et-xmlfile==1.0.1         # via openpyxl
factory-boy==2.12.0       # via -r requirements-dev.in
faker==4.0.1              # via factory-boy
fancycompleter==0.9.1     # via pdbpp
flake8==3.7.9             # via pytest-flake8
gunicorn==20.0.4          # via -r requirements.in
importlib-metadata==1.5.0  # via pluggy, pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via -r requirements-dev.in
jdcal==1.4.1              # via openpyxl
jedi==0.16.0              # via ipython
jmespath==0.9.5           # via boto3, botocore
markuppy==1.14            # via tablib
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
odfpy==1.4.1              # via tablib
openpyxl==3.0.3           # via tablib
packaging==20.3           # via pytest
parso==0.6.2              # via jedi
pathspec==0.7.0           # via black
pdbpp==0.10.2             # via -r requirements-dev.in
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.0.0             # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.4     # via ipython
psycopg2-binary==2.8.4    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via ipython, pdbpp
pyparsing==2.4.6          # via packaging
pyrepl==0.9.0             # via fancycompleter
pytest-cov==2.8.1         # via -r requirements-dev.in
pytest-django==3.8.0      # via -r requirements-dev.in
pytest-flake8==1.0.4      # via -r requirements-dev.in
pytest==5.3.5             # via -r requirements-dev.in, pytest-cov, pytest-django, pytest-flake8
python-dateutil==2.8.1    # via botocore, faker
pytz==2019.3              # via -r requirements.in, datetime, django
pyyaml==5.3               # via tablib
regex==2020.2.20          # via black
s3transfer==0.3.3         # via boto3
six==1.14.0               # via django-configurations, packaging, python-dateutil, traitlets
sqlparse==0.3.1           # via django
tablib[html,ods,xls,xlsx,yaml]==1.1.0  # via -r requirements.in, django-import-export
text-unidecode==1.3       # via faker
toml==0.10.0              # via black
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
urllib3==1.25.8           # via botocore
wcwidth==0.1.8            # via prompt-toolkit, pytest
whitenoise==5.0.1         # via -r requirements.in
wmctrl==0.3               # via pdbpp
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
zipp==3.1.0               # via importlib-metadata
zope.interface==4.7.2     # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
